+++ packages/playground/kit-demo-app/src/routes/todos/index.svelte
@@ -1,40 +1,9 @@
-<script context="module">
-	import { enhance } from '$lib/form';
-
-	// see https://kit.svelte.dev/docs#loading
-	export const load = async ({ fetch }) => {
-		const res = await fetch('/todos.json');
-
-		if (res.ok) {
-			const todos = await res.json();
-
-			return {
-				props: { todos }
-			};
-		}
-
-		const { message } = await res.json();
-
-		return {
-			error: new Error(message)
-		};
-	};
-</script>
-
 <script>
+	import { enhance } from '$lib/form';
 	import { scale } from 'svelte/transition';
 	import { flip } from 'svelte/animate';
 
 	export let todos;
-
-	async function patch(res) {
-		const todo = await res.json();
-
-		todos = todos.map((t) => {
-			if (t.uid === todo.uid) return todo;
-			return t;
-		});
-	}
 </script>
 
 <svelte:head>
@@ -46,13 +15,10 @@
 
 	<form
 		class="new"
-		action="/todos.json"
+		action="/todos"
 		method="post"
 		use:enhance={{
-			result: async (res, form) => {
-				const created = await res.json();
-				todos = [...todos, created];
-
+			result: async ({ form }) => {
 				form.reset();
 			}
 		}}
@@ -68,41 +34,34 @@
 			animate:flip={{ duration: 200 }}
 		>
 			<form
-				action="/todos/{todo.uid}.json?_method=patch"
+				action="/todos?_method=PATCH"
 				method="post"
 				use:enhance={{
-					pending: (data) => {
+					pending: ({ data }) => {
 						todo.done = !!data.get('done');
-					},
-					result: patch
+					}
 				}}
 			>
+				<input type="hidden" name="uid" value={todo.uid} />
 				<input type="hidden" name="done" value={todo.done ? '' : 'true'} />
 				<button class="toggle" aria-label="Mark todo as {todo.done ? 'not done' : 'done'}" />
 			</form>
 
-			<form
-				class="text"
-				action="/todos/{todo.uid}.json?_method=patch"
-				method="post"
-				use:enhance={{
-					result: patch
-				}}
-			>
+			<form class="text" action="/todos?_method=PATCH" method="post" use:enhance>
+				<input type="hidden" name="uid" value={todo.uid} />
 				<input aria-label="Edit todo" type="text" name="text" value={todo.text} />
 				<button class="save" aria-label="Save todo" />
 			</form>
 
 			<form
-				action="/todos/{todo.uid}.json?_method=delete"
+				action="/todos?_method=DELETE"
 				method="post"
 				use:enhance={{
-					result: () => {
-						todos = todos.filter((t) => t.uid !== todo.uid);
-					}
+					pending: () => (todo.pending_delete = true)
 				}}
 			>
-				<button class="delete" aria-label="Delete todo" />
+				<input type="hidden" name="uid" value={todo.uid} />
+				<button class="delete" aria-label="Delete todo" disabled={todo.pending_delete} />
 			</form>
 		</div>
 	{/each}