+++ packages/vite-plugin-svelte/src/utils/optimizer.ts
@@ -22,11 +22,10 @@ export async function handleOptimizeDeps(options: ResolvedOptions, viteConfig: R
 		return;
 	}
 
-	const viteMetadataPath = path.resolve(viteConfig.cacheDir, '_metadata.json');
+	const viteMetadataPath = findViteMetadataPath(viteConfig.cacheDir);
+	if (!viteMetadataPath) return;
 
-	if (!fs.existsSync(viteMetadataPath)) return;
-
-	const svelteMetadataPath = path.resolve(viteConfig.cacheDir, '_svelte_metadata.json');
+	const svelteMetadataPath = path.resolve(viteMetadataPath, '../_svelte_metadata.json');
 	const currentSvelteMetadata = JSON.stringify(generateSvelteMetadata(options), (_, value) => {
 		return typeof value === 'function' ? value.toString() : value;
 	});
@@ -46,4 +45,12 @@ function generateSvelteMetadata(options: ResolvedOptions) {
 		metadata[key] = options[key];
 	}
 	return metadata;
+}
+
+function findViteMetadataPath(cacheDir: string) {
+	const metadataPaths = ['_metadata.json', 'deps/_metadata.json'];
+	for (const metadataPath of metadataPaths) {
+		const viteMetadataPath = path.resolve(cacheDir, metadataPath);
+		if (fs.existsSync(viteMetadataPath)) return viteMetadataPath;
+	}
 }